Ability to execute external script when encountering connection issues due to permissions. Added safety to torrent cleanup so it doesn't delete download folder? Fixed missing error message on renames. Added option to find another magnet for a download in the interface.

+++ Traktor.Core/Services/Downloader/MediaDownloader.cs
@@ -374,7 +374,7 @@ namespace Traktor.Core.Services.Downloader
                 {
                     try
                     {
-                        if (Directory.Exists(torrentManager.SavePath))
+                        if (torrentManager.SavePath != this.DownloadPath && Directory.Exists(torrentManager.SavePath))
                             Directory.Delete(torrentManager.SavePath, true);
                         break;
                     }
